Conversational-Agent/requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain-openai>=0.0.5
langchain>=0.1.0
langchain-core>=0.1.0
neo4j>=5.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyttsx3>=2.90
soundfile>=0.12.0
sounddevice>=0.4.6 